backend/app/core/celery_app.py
from celery import Celery
import os
from dotenv import load_dotenv

load_dotenv()

celery_app = Celery(
    "micro_entrepreneur_app",
    broker=os.getenv("CELERY_BROKER_URL", "redis://localhost:6379/0"),
    backend=os.getenv("CELERY_RESULT_BACKEND", "redis://localhost:6379/0")
)

celery_app.conf.update(
    task_serializer="json",
    accept_content=["json"],
    result_serializer="json",
    timezone="UTC",
    enable_utc=True,
)

@celery_app.task
def send_follow_up_notification(customer_name: str, message: str):
    """
    Send a follow-up notification to a customer.
    In a real implementation, this would integrate with email/SMS/WhatsApp services.
    """
    # This is a placeholder - in reality, you would send the actual notification
    print(f"Sending follow-up to {customer_name}: {message}")
    return {"status": "sent", "customer": customer_name, "message": message}

@celery_app.task
def process_social_media_post(user_id: int, platform: str, content: str):
    """
    Process and publish a social media post.
    In a real implementation, this would integrate with social media APIs.
    """
    # This is a placeholder - in reality, you would post to the social platform
    print(f"Posting to {platform} for user {user_id}: {content}")
    return {"status": "posted", "platform": platform, "user_id": user_id}