backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.3
httpx==0.25.1
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
alembic==1.12.1
slowapi==0.1.9
bleach==6.1.0
google-generativeai>=0.3.0
Pillow>=10.0.0uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
httpx==0.25.1
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
alembic==1.12.1
slowapi==0.1.9
bleach==6.1.0